+++ Assets/Tests/InputSystem/Plugins/UITests.cs
@@ -3,15 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using NUnit.Framework;
 using UnityEngine;
-using UnityEngine.Scripting;
 using UnityEngine.EventSystems;
 using UnityEngine.InputSystem;
-using UnityEngine.InputSystem.Controls;
-using UnityEngine.InputSystem.Layouts;
 using UnityEngine.InputSystem.LowLevel;
 using UnityEngine.InputSystem.UI;
-using UnityEngine.InputSystem.Utilities;
 using UnityEngine.TestTools;
+using UnityEngine.TestTools.Utils;
 using UnityEngine.UI;
 using TouchPhase = UnityEngine.InputSystem.TouchPhase;
 
@@ -113,8 +110,8 @@ internal class UITests : InputTestFixture
     }
 
     [UnityTest]
-    [Category("Actions")]
-    public IEnumerator Actions_CanDriveUIFromMouse()
+    [Category("UI")]
+    public IEnumerator UI_CanDriveUIFromMouse()
     {
         // Create devices.
         var mouse = InputSystem.AddDevice<Mouse>();
@@ -254,8 +251,8 @@ internal class UITests : InputTestFixture
     }
 
     [UnityTest]
-    [Category("Actions")]
-    public IEnumerator Actions_TouchActionsCanDriveUIAndDistinguishMultipleTouches()
+    [Category("UI")]
+    public IEnumerator UI_TouchActionsCanDriveUIAndDistinguishMultipleTouches()
     {
         // Create devices.
         var touchScreen = InputSystem.AddDevice<Touchscreen>();
@@ -441,10 +438,10 @@ internal class UITests : InputTestFixture
     }
 
     [UnityTest]
-    [Category("Actions")]
+    [Category("UI")]
     // Check that two players can have separate UI, and that both selections will stay active when
     // clicking on UI with the mouse, using MultiPlayerEventSystem.playerRoot to match UI to the players.
-    public IEnumerator Actions_CanOperateMultiplayerUIGloballyUsingMouse()
+    public IEnumerator UI_CanOperateMultiplayerUIGloballyUsingMouse()
     {
         var mouse = InputSystem.AddDevice<Mouse>();
 
@@ -523,10 +520,10 @@ internal class UITests : InputTestFixture
     }
 
     [UnityTest]
-    [Category("Actions")]
+    [Category("UI")]
     // Check that two players can have separate UI and control it using separate gamepads, using
     // MultiplayerEventSystem.
-    public IEnumerator Actions_CanOperateMultiplayerUILocallyUsingGamepads()
+    public IEnumerator UI_CanOperateMultiplayerUILocallyUsingGamepads()
     {
         // Create devices.
         var gamepads = new[] { InputSystem.AddDevice<Gamepad>(), InputSystem.AddDevice<Gamepad>() };
@@ -638,8 +635,8 @@ internal class UITests : InputTestFixture
     }
 
     [UnityTest]
-    [Category("Actions")]
-    public IEnumerator Actions_CanDriveUIFromGamepad()
+    [Category("UI")]
+    public IEnumerator UI_CanDriveUIFromGamepad()
     {
         // Create devices.
         var gamepad = InputSystem.AddDevice<Gamepad>();
@@ -741,8 +738,8 @@ internal class UITests : InputTestFixture
     }
 
     [Test]
-    [Category("Actions")]
-    public void Actions_CanReassignUIActions()
+    [Category("UI")]
+    public void UI_CanReassignUIActions()
     {
         var go = new GameObject();
         go.AddComponent<EventSystem>();
@@ -789,6 +786,215 @@ internal class UITests : InputTestFixture
         Assert.That(uiModule.point?.action, Is.Null);
     }
 
+    [Test]
+    [Category("UI")]
+    public void UI_CanDriveVirtualMouseCursorFromGamepad()
+    {
+        const float kCursorSpeed = 100;
+        const float kScrollSpeed = 25;
+
+        var eventSystemGO = new GameObject();
+        eventSystemGO.SetActive(false);
+        eventSystemGO.AddComponent<EventSystem>();
+        eventSystemGO.AddComponent<InputSystemUIInputModule>();
+
+        var canvasGO = new GameObject();
+        canvasGO.SetActive(false);
+        var canvas = canvasGO.AddComponent<Canvas>();
+
+        var cursorGO = new GameObject();
+        cursorGO.SetActive(false);
+        var cursorTransform = cursorGO.AddComponent<RectTransform>();
+        var cursorInput = cursorGO.AddComponent<VirtualMouseInput>();
+        cursorInput.cursorSpeed = kCursorSpeed;
+        cursorInput.scrollSpeed = kScrollSpeed;
+        cursorInput.cursorTransform = cursorTransform;
+        cursorInput.canvas = canvas;
+        cursorTransform.SetParent(canvasGO.transform, worldPositionStays: false);
+        cursorTransform.pivot = new Vector2(0.5f, 0.5f);
+        cursorTransform.anchorMin = Vector2.zero;
+        cursorTransform.anchorMax = Vector2.zero;
+        cursorTransform.anchoredPosition = new Vector2(123, 234);
+
+        var positionAction = new InputAction(type: InputActionType.Value, binding: "<Gamepad>/*stick");
+        var leftButtonAction = new InputAction(binding: "<Gamepad>/buttonSouth");
+        var rightButtonAction = new InputAction(binding: "<Gamepad>/rightShoulder");
+        var middleButtonAction = new InputAction(binding: "<Gamepad>/leftShoulder");
+        var forwardButtonAction = new InputAction(binding: "<Gamepad>/buttonWest");
+        var backButtonAction = new InputAction(binding: "<Gamepad>/buttonEast");
+        var scrollWheelAction = new InputAction();
+        scrollWheelAction.AddCompositeBinding("2DVector")
+            .With("Up", "<Gamepad>/leftTrigger")
+            .With("Down", "<Gamepad>/rightTrigger")
+            .With("Left", "<Gamepad>/dpad/left")
+            .With("Right", "<Gamepad>/dpad/right");
+
+        cursorInput.stickAction = new InputActionProperty(positionAction);
+        cursorInput.leftButtonAction = new InputActionProperty(leftButtonAction);
+        cursorInput.rightButtonAction = new InputActionProperty(rightButtonAction);
+        cursorInput.middleButtonAction = new InputActionProperty(middleButtonAction);
+        cursorInput.scrollWheelAction = new InputActionProperty(scrollWheelAction);
+        cursorInput.forwardButtonAction = new InputActionProperty(forwardButtonAction);
+        cursorInput.backButtonAction = new InputActionProperty(backButtonAction);
+
+        var gamepad = InputSystem.AddDevice<Gamepad>();
+
+        // Get rid of deadzones to simplify computations.
+        InputSystem.settings.defaultDeadzoneMin = 0;
+        InputSystem.settings.defaultDeadzoneMax = 1;
+
+        eventSystemGO.SetActive(true);
+        canvasGO.SetActive(true);
+        cursorGO.SetActive(true);
+
+        // Make sure the component added a virtual mouse.
+        var virtualMouse = Mouse.current;
+        Assert.That(virtualMouse, Is.Not.Null);
+        Assert.That(virtualMouse.layout, Is.EqualTo("VirtualMouse"));
+        Assert.That(cursorInput.virtualMouse, Is.SameAs(virtualMouse));
+
+        // Make sure we can disable and re-enable the component.
+        cursorGO.SetActive(false);
+
+        Assert.That(Mouse.current, Is.Null);
+
+        cursorGO.SetActive(true);
+
+        Assert.That(Mouse.current, Is.Not.Null);
+        Assert.That(Mouse.current, Is.SameAs(virtualMouse));
+
+        // Ensure everything is at default values.
+        // Starting position should be that of the cursor's initial transform.
+        Assert.That(virtualMouse.position.ReadValue(), Is.EqualTo(new Vector2(123, 234)).Using(Vector2EqualityComparer.Instance));
+        Assert.That(virtualMouse.delta.ReadValue(), Is.EqualTo(Vector2.zero));
+        Assert.That(virtualMouse.scroll.ReadValue(), Is.EqualTo(Vector2.zero));
+        Assert.That(virtualMouse.leftButton.isPressed, Is.False);
+        Assert.That(virtualMouse.rightButton.isPressed, Is.False);
+        Assert.That(virtualMouse.middleButton.isPressed, Is.False);
+        Assert.That(cursorTransform.anchoredPosition, Is.EqualTo(new Vector2(123, 234)));
+
+        // Now move the mouse cursor with the left stick and ensure we get a response.
+        currentTime = 1;
+        Set(gamepad.leftStick, new Vector2(0.25f, 0.75f));
+
+        // No time has passed yet so first frame shouldn't move at all.
+        Assert.That(virtualMouse.position.ReadValue(), Is.EqualTo(new Vector2(123, 234)).Using(Vector2EqualityComparer.Instance));
+        Assert.That(virtualMouse.delta.ReadValue(), Is.EqualTo(Vector2.zero));
+        Assert.That(cursorTransform.anchoredPosition, Is.EqualTo(new Vector2(123, 234)));
+
+        currentTime = 1.4;
+        InputSystem.Update();
+
+        const float kFirstDeltaX = kCursorSpeed * 0.25f * 0.4f;
+        const float kFirstDeltaY = kCursorSpeed * 0.75f * 0.4f;
+
+        Assert.That(virtualMouse.position.ReadValue(), Is.EqualTo(new Vector2(123 + kFirstDeltaX, 234 + kFirstDeltaY)).Using(Vector2EqualityComparer.Instance));
+        Assert.That(virtualMouse.delta.ReadValue(), Is.EqualTo(new Vector2(kFirstDeltaX, kFirstDeltaY)).Using(Vector2EqualityComparer.Instance));
+        Assert.That(cursorTransform.anchoredPosition, Is.EqualTo(new Vector2(123 + kFirstDeltaX, 234 + kFirstDeltaY)).Using(Vector2EqualityComparer.Instance));
+
+        // Each update should move the cursor along while the stick is actuated.
+        currentTime = 2;
+        InputSystem.Update();
+
+        const float kSecondDeltaX = kCursorSpeed * 0.25f * 0.6f;
+        const float kSecondDeltaY = kCursorSpeed * 0.75f * 0.6f;
+
+        Assert.That(virtualMouse.position.ReadValue(), Is.EqualTo(new Vector2(123 + kFirstDeltaX + kSecondDeltaX, 234 + kFirstDeltaY + kSecondDeltaY)).Using(Vector2EqualityComparer.Instance));
+        Assert.That(virtualMouse.delta.ReadValue(), Is.EqualTo(new Vector2(kSecondDeltaX, kSecondDeltaY)).Using(Vector2EqualityComparer.Instance));
+        Assert.That(cursorTransform.anchoredPosition, Is.EqualTo(new Vector2(123 + kFirstDeltaX + kSecondDeltaX, 234 + kFirstDeltaY + kSecondDeltaY)).Using(Vector2EqualityComparer.Instance));
+
+        // Only the final state of the stick in an update should matter.
+        currentTime = 3;
+        InputSystem.QueueStateEvent(gamepad, new GamepadState { leftStick = new Vector2(0.34f, 0.45f)});
+        InputSystem.QueueStateEvent(gamepad, new GamepadState { leftStick = new Vector2(0.45f, 0.56f)});
+        InputSystem.Update();
+
+        const float kThirdDeltaX = kCursorSpeed * 0.45f;
+        const float kThirdDeltaY = kCursorSpeed * 0.56f;
+
+        Assert.That(virtualMouse.position.ReadValue(), Is.EqualTo(new Vector2(123 + kFirstDeltaX + kSecondDeltaX + kThirdDeltaX, 234 + kFirstDeltaY + kSecondDeltaY + kThirdDeltaY)).Using(Vector2EqualityComparer.Instance));
+        Assert.That(virtualMouse.delta.ReadValue(), Is.EqualTo(new Vector2(kThirdDeltaX, kThirdDeltaY)).Using(Vector2EqualityComparer.Instance));
+        Assert.That(cursorTransform.anchoredPosition, Is.EqualTo(new Vector2(123 + kFirstDeltaX + kSecondDeltaX + kThirdDeltaX, 234 + kFirstDeltaY + kSecondDeltaY + kThirdDeltaY)).Using(Vector2EqualityComparer.Instance));
+
+        var leftClickAction = new InputAction(binding: "<Mouse>/leftButton");
+        var middleClickAction = new InputAction(binding: "<Mouse>/middleButton");
+        var rightClickAction = new InputAction(binding: "<Mouse>/rightButton");
+        var forwardClickAction = new InputAction(binding: "<Mouse>/forwardButton");
+        var backClickAction = new InputAction(binding: "<Mouse>/backButton");
+        var scrollAction = new InputAction(binding: "<Mouse>/scroll");
+
+        leftClickAction.Enable();
+        middleClickAction.Enable();
+        rightClickAction.Enable();
+        forwardClickAction.Enable();
+        backClickAction.Enable();
+        scrollAction.Enable();
+
+        // Press buttons.
+        PressAndRelease(gamepad.buttonSouth);
+        Assert.That(leftClickAction.triggered);
+        PressAndRelease(gamepad.rightShoulder);
+        Assert.That(rightClickAction.triggered);
+        PressAndRelease(gamepad.leftShoulder);
+        Assert.That(middleClickAction.triggered);
+        PressAndRelease(gamepad.buttonWest);
+        Assert.That(forwardClickAction.triggered);
+        PressAndRelease(gamepad.buttonEast);
+        Assert.That(backClickAction.triggered);
+
+        // Scroll wheel.
+        Set(gamepad.leftTrigger, 0.5f);
+        Assert.That(scrollAction.ReadValue<Vector2>(), Is.EqualTo(new Vector2(0, kScrollSpeed * 0.5f)).Using(Vector2EqualityComparer.Instance));
+        Set(gamepad.rightTrigger, 0.3f);
+        Assert.That(scrollAction.ReadValue<Vector2>(), Is.EqualTo(new Vector2(0, kScrollSpeed * (0.5f - 0.3f))).Using(Vector2EqualityComparer.Instance));
+        Set(gamepad.leftTrigger, 0);
+        Assert.That(scrollAction.ReadValue<Vector2>(), Is.EqualTo(new Vector2(0, -kScrollSpeed * 0.3f)).Using(Vector2EqualityComparer.Instance));
+        Press(gamepad.dpad.left);
+        Assert.That(scrollAction.ReadValue<Vector2>(), Is.EqualTo(new Vector2(-kScrollSpeed, -kScrollSpeed * 0.3f)).Using(Vector2EqualityComparer.Instance));
+        Press(gamepad.dpad.right);
+        Assert.That(scrollAction.ReadValue<Vector2>(), Is.EqualTo(new Vector2(0, -kScrollSpeed * 0.3f)).Using(Vector2EqualityComparer.Instance));
+        Release(gamepad.dpad.left);
+        Assert.That(scrollAction.ReadValue<Vector2>(), Is.EqualTo(new Vector2(kScrollSpeed, -kScrollSpeed * 0.3f)).Using(Vector2EqualityComparer.Instance));
+    }
+
+    // Right now, text input in uGUI is picked up from IMGUI events. ATM they're still out of reach for us.
+    // Hopefully something we can solve as part of getting rid of the old input system.
+    [Test]
+    [Category("UI")]
+    [Ignore("TODO")]
+    public void TODO_UI_CanDriveTextInput()
+    {
+        var keyboard = InputSystem.AddDevice<Keyboard>();
+
+        var eventSystemGO = new GameObject();
+        eventSystemGO.AddComponent<EventSystem>();
+        eventSystemGO.AddComponent<InputSystemUIInputModule>();
+
+        var canvasGO = new GameObject();
+        canvasGO.AddComponent<Canvas>();
+
+        var inputFieldGO = new GameObject();
+        inputFieldGO.transform.SetParent(canvasGO.transform);
+        var inputField = inputFieldGO.AddComponent<InputField>();
+        inputField.text = string.Empty;
+
+        InputSystem.QueueTextEvent(keyboard, 'a');
+        InputSystem.QueueTextEvent(keyboard, 'b');
+        InputSystem.QueueTextEvent(keyboard, 'c');
+        InputSystem.Update();
+
+        Assert.That(inputField.text, Is.EqualTo("abc"));
+    }
+
+    ////TODO: We need to override BaseInput which currently is still hooked to the old input system APIs.
+    [Test]
+    [Category("UI")]
+    [Ignore("TODO")]
+    public void TODO_UI_CanDriveIME()
+    {
+        Assert.Fail();
+    }
+
 // The tracked device tests fail with NullReferenceException in the windows editor on yamato. I cannot reproduce this locally, so will disable them on windows for now.
 #if !UNITY_EDITOR_WIN
 
@@ -824,8 +1030,8 @@ internal class UITests : InputTestFixture
     }
 
     [UnityTest]
-    [Category("Actions")]
-    public IEnumerator Actions_CanDriveUIFromTrackedDevice()
+    [Category("UI")]
+    public IEnumerator UI_CanDriveUIFromTrackedDevice()
     {
         // Create device.
         InputSystem.RegisterLayout<TestTrackedDevice>();
@@ -958,8 +1164,8 @@ internal class UITests : InputTestFixture
     }
 
     [UnityTest]
-    [Category("Actions")]
-    public IEnumerator Actions_CanDriveUIFromMultipleTrackedDevices()
+    [Category("UI")]
+    public IEnumerator UI_CanDriveUIFromMultipleTrackedDevices()
     {
         // Create device.
         InputSystem.RegisterLayout<TestTrackedDevice>();