+++ Packages/com.unity.inputsystem/InputSystem/InputSystem.cs
@@ -2889,6 +2889,8 @@ namespace UnityEngine.InputSystem
 #if !UNITY_DISABLE_DEFAULT_INPUT_PLUGIN_INITIALIZATION
         private static void PerformDefaultPluginInitialization()
         {
+            UISupport.Initialize();
+
             #if UNITY_EDITOR || UNITY_STANDALONE || UNITY_WSA || UNITY_IOS
             XInputSupport.Initialize();
             #endif